Refactor Metrics models
