+++ src/shared/models.ts
@@ -4,27 +4,6 @@ import {
     SHARED_Model__RepoWithCommitCountsAndLanguages,
 } from './models/models/Repos'
 
-/* ======== metrics ======== */
-
-export interface SHARED_Data__WeeklyContributionCommitActivityData {
-    w: number
-    a: number
-    d: number
-    c: number
-}
-
-export interface SHARED_Data__ContributorCommitActivity {
-    author: {
-        login: string
-    }
-    total: number
-    weeks: SHARED_Data__WeeklyContributionCommitActivityData[]
-}
-
-export interface SHARED_APIFields__GetContributorCommitActivity extends SHARED_APIFields__BASE {
-    activity?: SHARED_Data__ContributorCommitActivity
-}
-
 /* ======== languages ======== */
 
 export interface SHARED_Model__AllLanguageStats {
